+++ lib/iex/api.rb
@@ -1,4 +1,5 @@
 require_relative 'endpoints/advanced_stats'
+require_relative 'endpoints/balance_sheet'
 require_relative 'endpoints/chart'
 require_relative 'endpoints/company'
 require_relative 'endpoints/dividends'